+++ CHANGELOG.md
@@ -1,13 +1,18 @@
-# BetterBlizzFrames 2.0.5b
+# BetterBlizzFrames 2.0.6
 ## Midnight
 ### New
+- Every filter setting for Target/FocusFrame now has a "Friendly/Enemy" toggle box. So you can for example show only your buffs on Friendly Targets but all buffs on Enemy Targets.
+- New "Dispellable" filter for debuffs on Target/Focus. 
 - New Ceit profile (www.twitch.tv/ceitxd). Thank you for sharing!
 ### Tweak
-- Finally fix "Mirror TargetFrame" setting for Midnight.
-- Fix "Purgeable" filter to not rely on whether you can dispel or not but just the nature of the buff if its dispellable or not.
+- Finally fix "Mirror TargetFrame" setting for Midnight. Makes the Player Portrait round; a mirrored version of TargetFrame.
 - Add a tiny crop on spec icons for spec icon portraits setting (to hide borders sticking out)
 - Tiny tweak to ocd setting and level position.
+### Bugfix
+- Fix "Purgeable" filter to not rely on whether you can dispel or not but just the nature of the buff if its dispellable or not.
 - Fix castbars flashing white on some castbar color settings at the end of cast.
+- Fix a spacing issue between Target/FocusFrame and auras with aura settings enabled and some specific filters.
+- Fix an issue with glows on auras not showing on friendly units.
 
 # BetterBlizzFrames 2.0.5
 ## Midnight